add site.lhs
This should eventually just replace site.hs.  The goal is to embed my
planned Hakyll tutorial directly into the Hakyll configuration file,
using literate Haskell. btw, site.lhs already compiles and works, with
just a few problems (which are also present in site.hs, so that's not
really an issue), like bird tracks showing up in the HTML representation
of the site.